% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{buttonBlock}
\alias{buttonBlock}
\title{Create a block-level button container}
\usage{
buttonBlock(..., border = "top")
}
\arguments{
\item{...}{One or more \code{\link[=shiny]{actionButton}} or
\code{\link[=shiny]{downloadButton}} objects.}

\item{border}{Zero or more of \code{c("top", "bottom")}, indicating which
sides should have borders, if any.}
}
\description{
Creates a full-width container for one or more buttons. The horizontal space
will be evenly divided among any buttons that are added.
}
\details{
When using \code{buttonBlock} with a \code{tabstripPanel}, consider passing
the \code{buttonBlock} to \code{tabstripPanel} as the \code{between}
argument.
}
\examples{
library(shiny)

buttonBlock(
  actionButton("reset", "Reset to defaults"),
  actionButton("clear", "Clear all")
)

}

