% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{gadgetPage}
\alias{gadgetPage}
\title{Page function for Shiny Gadgets}
\usage{
gadgetPage(..., theme = NULL)
}
\arguments{
\item{...}{Elements to include within the page.}
}
\description{
Designed to serve as the outermost function call for your gadget UI. Similar
to \code{\link{fillPage}}, but always includes the Bootstrap CSS library, and
is designed to contain \code{\link{titlebar}}, \code{\link{tabstripPanel}},
\code{\link{contentPanel}}, etc.
}

