% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{miniTabstripPanel}
\alias{miniTabstripPanel}
\title{Create a tabstrip panel}
\usage{
miniTabstripPanel(..., id = NULL, selected = NULL, between = NULL)
}
\arguments{
\item{...}{\code{\link[=shiny]{tabPanel}} elements to include in the tabset.}

\item{id}{If provided, you can use \code{input$}\emph{\code{id}} in your
server logic to determine which of the current tabs is active. The value
will correspond to the \code{value} argument that is passed to
\code{\link{tabPanel}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{between}{A tag or list of tags that should be inserted between the
content (above) and tabstrip (below).}
}
\description{
Create a tabstrip panel that contains \code{\link[=shiny]{tabPanel}}
elements. Similar to \code{\link[=shiny]{tabsetPanel}}, but optimized for
small page sizes like mobile devices or the RStudio Viewer pane.
}
\examples{
library(shiny)

miniTabstripPanel(
  tabPanel("Data",
    selectInput("dataset", "Data set", ls("package:datasets"))),
  tabPanel("Subset",
    uiOutput("subset_ui")
  )
)

}

