% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{titlebar}
\alias{titlebar}
\alias{titlebarButton}
\title{Create a title bar}
\usage{
titlebar(title, left = NULL, right = titlebarButton("done", "Done", primary
  = TRUE))

titlebarButton(inputId, label, primary = FALSE)
}
\arguments{
\item{title}{The title of the gadget. If this needs to be dynamic, pass
\code{\link[=shiny]{textOutput}} with \code{inline = TRUE}.}

\item{left}{The \code{titlebarButton} to put on the left, or \code{NULL} for
none.}

\item{right}{The \code{titlebarButton} to put on the right, or \code{NULL}
for none. Defaults to a primary "Done" button that can be handled using
\code{observeEvent(input$done, \{...\})}.}

\item{inputId}{The \code{input} slot that will be used to access the button.}

\item{label}{The text label to display on the button.}

\item{primary}{If \code{TRUE}, render the button in a bold color to indicate
that it is the primary action of the gadget.}
}
\description{
Creates a title bar for a Shiny Gadget. Intended to be used with
\code{\link{gadgetPage}}. Title bars contain a title, and optionally, a
\code{titlebarButton} on the left and/or right sides.
}

